重構DOMHelper、重構addListener、整理logging、bug處理
