+++ .vscode/settings.json
@@ -1,9 +1,14 @@
 {
     "cSpell.words": [
+        "assjs",
+        "btns",
+        "Meiryo",
+        "nexttrack",
         "Onedrive",
         "playlistexclude",
         "playlistinclude",
         "startplaylist",
-        "videojs"
+        "videojs",
+        "webvtt"
     ]
 }
