+++ src/Models/Message.ts
@@ -1,11 +1,12 @@
 export class Message<T> implements IMessage<T> {
     Name = '';
     Data: T;
-    constructor(name: string, data: T) {
+    constructor(name: string, data?: T) {
         this.Name = name;
-        this.Data = data;
+        this.Data = data ?? {} as T;
     }
 }
+
 export interface IMessage<T> {
     Name: string;
     Data: T;
