+++ src/background.ts
@@ -1,16 +1,22 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { IPlaylist } from './Models/IPlaylist';
 import { IMessage } from './Models/Message';
-import * as UrlHelper from './UrlHelper';
-import { urlParams, url } from './UrlHelper';
+import * as UrlHelper from './Helper/URLHelper';
+import { urlParams, url } from './Helper/URLHelper';
 
 let DisabledPlaylists: string[] = [];
-let shuffle = false;
 let shuffleList: number[] = [];
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 let myPlaylist: any[][] = [];
 let Playlists: IPlaylist[] = [];
+
+// Clear storage
+chrome.storage.local.remove(['shuffleList', 'myPlaylist', 'params']);
+
+addListeners();
+fetchPlaylists();
+
 
 async function fetchPlaylists(): Promise<unknown> {
     const response = await fetch('https://github.com/jim60105/Playlists/raw/minify/Playlists.jsonc');
@@ -18,8 +24,75 @@ async function fetchPlaylists(): Promise<unknown> {
     return Playlists = json;
 }
 
+function addListeners() {
+    function _addListener<T>(_name: string, callback: (message: IMessage<T>, sender: chrome.runtime.MessageSender, sendResponse: (response?: unknown) => void) => void) {
+        chrome.runtime.onMessage.addListener(function (message: IMessage<T>, sender, sendResponse) {
+            if (!message || message.Name !== _name) return;
+            console.debug('Message received from Content Script: %o', message);
+
+            callback(message, sender, sendResponse);
+            return true;
+        });
+    }
+
+    _addListener<string>(
+        'LoadPlaylists',
+        async (message, sender, sendResponse) => {
+            await UrlHelper.prepareUrlParams(message.Data);
+            await LoadPlayLists();
+            sendResponse();
+        });
+    _addListener<string>(
+        'CheckList',
+        (message, sender, sendResponse) => {
+            sendResponse(CheckList());
+        });
+    _addListener<{ 'index': number, 'UIClick': boolean; }>(
+        'NextSongToBackground',
+        (message, sender, sendResponse) => {
+            if (sender.tab?.id)
+                NextSong(message.Data.index, sender.tab.id, message.Data.UIClick);
+            sendResponse();
+        });
+    _addListener<boolean>(
+        'StepShuffle',
+        (message, sender, sendResponse) => {
+            shuffleList.push(shuffleList.shift() ?? 0);
+            chrome.storage.local.set({ 'shuffleList': shuffleList });
+            if (sender.tab?.id)
+                NextSong(shuffleList[0], sender.tab.id);
+            sendResponse();
+        });
+    _addListener<boolean>(
+        'GetNowPlaying',
+        (message, sender, sendResponse) => {
+            sendResponse(myPlaylist[CheckList()]);
+        });
+}
+
 async function LoadPlayLists() {
     const LoadedPlaylists = new Map();
+    await getStorageLists();
+
+    console.groupCollapsed('Playlists');
+    console.table(Playlists);
+    console.groupEnd();
+
+    console.groupCollapsed('Loaded Playlists');
+    await LoadAllPlaylists();
+    await ConcatPlaylistsIntoMyPlaylists();
+    console.groupEnd();
+
+    console.groupCollapsed('myPlaylists');
+    console.table(myPlaylist);
+    console.groupEnd();
+
+    const shuffle = urlParams.has('shuffle') && urlParams.get('shuffle') !== '0';
+    console.log('Shuffle: %o', shuffle);
+    console.log('StartPlayList: ', urlParams.has('startplaylist'));
+    await MakeNewShuffleList();
+    return Promise.resolve(true);
+
 
     function getStorageLists(): Promise<unknown[]> {
         const promises = [
@@ -78,7 +151,8 @@ async function LoadPlayLists() {
                     })
                     .then(_playlist => {
                         LoadedPlaylists.set(listName, _playlist);
-                        console.log('Load %s: %o', listName, _playlist);
+                        console.log('Load %s', listName);
+                        console.table(_playlist);
                     });
             }
         } else {
@@ -100,62 +174,52 @@ async function LoadPlayLists() {
         return Promise.all(promises);
     }
 
-    function ConcatPlaylistsIntoMyPlaylists() {
+    async function ConcatPlaylistsIntoMyPlaylists() {
         myPlaylist = [];
+        // It is important to load here in the order of 'Playlists'.
         Playlists.forEach((playlist) => {
             if (LoadedPlaylists.has(playlist.name)) {
                 myPlaylist = myPlaylist.concat(LoadedPlaylists.get(playlist.name));
             }
         });
-        chrome.storage.local.set({ 'myPlaylist': myPlaylist });
+
+        await chrome.storage.local.set({ 'myPlaylist': myPlaylist });
     }
 
-    function MakeShuffleList(length: number) {
-        const shuffleList = [];
-        for (let i = 0; i < length; ++i) shuffleList[i] = i;
+    function MakeRandomArray(length: number) {
+        const array = [];
+        for (let i = 0; i < length; ++i) array[i] = i;
 
         // http://stackoverflow.com/questions/962802#962890
         let tmp,
             current,
-            top = shuffleList.length;
+            top = array.length;
         if (top)
             while (--top) {
                 current = Math.floor(Math.random() * (top + 1));
-                tmp = shuffleList[current];
-                shuffleList[current] = shuffleList[top];
-                shuffleList[top] = tmp;
+                tmp = array[current];
+                array[current] = array[top];
+                array[top] = tmp;
             }
 
-        console.log('Make new shuffleList');
-        return shuffleList;
+        return array;
     }
 
-    await getStorageLists();
-
-    console.log(Playlists);
-    await LoadAllPlaylists();
-    ConcatPlaylistsIntoMyPlaylists();
-
-    shuffle = urlParams.has('shuffle') && urlParams.get('shuffle') !== '0';
-    chrome.storage.local.set({ 'shuffle': shuffle });
-    console.log('Shuffle: %o', shuffle);
-
-    // Make shuffle list when length not match, or starting playlist
-    if (shuffle
-        && (urlParams.has('startplaylist')
-            || shuffleList.length != myPlaylist.length)) {
-        shuffleList = MakeShuffleList(myPlaylist.length);
-
-        chrome.storage.local.set({ 'shuffleList': shuffleList });
+    async function MakeNewShuffleList() {
+        if (shuffle
+            && (shuffleList.length !== myPlaylist.length
+                || urlParams.has('startplaylist'))) {
+            console.log('Making new shuffleList...');
+            shuffleList = MakeRandomArray(myPlaylist.length);
+            await chrome.storage.local.set({ 'shuffleList': shuffleList });
+        }
     }
-
 }
 
 function CheckList(): number {
     UrlHelper.CleanUpParameters();
 
     /**
-     * ["VideoID", StartTime, EndTime, "Title"]
      * VideoID: 必須用引號包住，為字串型態。
      * StartTime: 只能是非負數。如果要從頭播放，輸入0
      * EndTime: 只能是非負數。如果要播放至尾，輸入0
@@ -203,10 +267,10 @@ function CheckList(): number {
     }
     if (flag) {
         console.log(`Playing on Playlist No.${i}`);
-        console.log(`Name : ${myPlaylist[i][3]}`);
-        console.log(`URL  : ${myPlaylist[i][0]}`);
-        console.log(`Start: ${myPlaylist[i][1]}`);
-        console.log(`End  : ${myPlaylist[i][2]}`);
+        console.debug(`Name : ${myPlaylist[i][3]}`);
+        console.debug(`URL  : ${myPlaylist[i][0]}`);
+        console.debug(`Start: ${myPlaylist[i][1]}`);
+        console.debug(`End  : ${myPlaylist[i][2]}`);
     } else {
         console.log('Not playing in the playlist.');
         i = -1;
@@ -215,6 +279,7 @@ function CheckList(): number {
 }
 
 async function NextSong(index: number, tabId: number, UIClick = false) {
+    const shuffle = urlParams.has('shuffle') && urlParams.get('shuffle') !== '0';
     if (myPlaylist.length == 0) {
         console.warn('Playlist not loaded! Reloading playlists...');
         await LoadPlayLists();
@@ -226,7 +291,7 @@ async function NextSong(index: number, tabId: number, UIClick = false) {
         index = 0;
     }
 
-    // // Send "next song" outside the iframe
+    // // Send 'next song' outside the iframe
     // if ('/embed/' == window.location.pathname) {
     //     if (!UIClick) {
     //         parent.postMessage('song end', '*');
@@ -257,6 +322,7 @@ async function NextSong(index: number, tabId: number, UIClick = false) {
     urlParams.set('t', nextSong[1]);
     urlParams.set('end', nextSong[2]);
 
+    let newURL: string;
     if (nextSong[0].indexOf('http') >= 0) {
         // URL (Onedrive)
         const _url = new URL(nextSong[0]);
@@ -265,79 +331,19 @@ async function NextSong(index: number, tabId: number, UIClick = false) {
                 urlParams.set(key, value);
             });
         }
-        await UrlHelper.SaveToStorage();
-        chrome.tabs.update(tabId, { url: `${_url.origin}${_url.pathname}?${urlParams.toString()}${_url.hash}` });
+        newURL = `${_url.origin}${_url.pathname}?${urlParams.toString()}${_url.hash}`;
     } else {
-        await UrlHelper.SaveToStorage();
         // ID
         if (nextSong[0].length > 20) {
             // Google Drive
-            chrome.tabs.update(tabId, { url: `https://drive.google.com/file/d/${nextSong[0]}/view?${urlParams.toString()}` });
+            newURL = `https://drive.google.com/file/d/${nextSong[0]}/view?${urlParams.toString()}`;
         } else {
             // Youtube
-            chrome.tabs.update(tabId, { url: `https://www.youtube.com/watch?${urlParams.toString()}` });
+            newURL = `https://www.youtube.com/watch?${urlParams.toString()}`;
         }
     }
+    await UrlHelper.SaveToStorage();
+
+    console.log('Redirect: %s', newURL);
+    chrome.tabs.update(tabId, { url: newURL });
 }
-
-function addListeners() {
-    function addLoadPlaylistsListener() {
-        chrome.runtime.onMessage.addListener(async function (message: IMessage<string>, sender, sendResponse) {
-            if (!message || message.Name !== 'LoadPlaylists')
-                return;
-
-            console.debug('Message received from contentScript.js: %o', message);
-            await UrlHelper.prepareUrlParams(message.Data);
-
-            await LoadPlayLists();
-            sendResponse(true);
-        });
-    }
-    function addCheckListListener() {
-        chrome.runtime.onMessage.addListener(function (message: IMessage<unknown>, sender, sendResponse) {
-            if (message && message.Name === 'CheckList') {
-                sendResponse(CheckList());
-            }
-        });
-    }
-
-    function addNextSongListener() {
-        chrome.runtime.onMessage.addListener(async function (message: IMessage<{ 'index': number, 'UIClick': boolean; }>, sender, sendResponse) {
-            if (message && message.Name === 'NextSongToBackground'
-                && sender.tab?.id) {
-                await NextSong(message.Data.index, sender.tab.id, message.Data.UIClick);
-            }
-        });
-    }
-
-    function addStepShuffleListener() {
-        chrome.runtime.onMessage.addListener(async function (message: IMessage<boolean>, sender, sendResponse) {
-            if (message && message.Name === 'StepShuffle'
-                && sender.tab?.id) {
-                shuffleList.push(shuffleList.shift() ?? 0);
-                chrome.storage.local.set({ 'shuffleList': shuffleList });
-                NextSong(shuffleList[0], sender.tab.id);
-            }
-        });
-    }
-
-    function addNowPlayingListener() {
-        chrome.runtime.onMessage.addListener(async function (message: IMessage<boolean>, sender, sendResponse) {
-            if (message && message.Name === 'GetNowPlaying') {
-                sendResponse(myPlaylist[CheckList()]);
-            }
-        });
-    }
-
-    addLoadPlaylistsListener();
-    addNextSongListener();
-    addCheckListListener();
-    addStepShuffleListener();
-    addNowPlayingListener();
-}
-
-// Clear urlParams
-UrlHelper.SaveToStorage();
-
-addListeners();
-fetchPlaylists();
