+++ tsconfig.json
@@ -2,7 +2,7 @@
     "compilerOptions": {
        "strict": true,
        "module": "commonjs",
-       "target": "es6",
+       "target": "es2017",
        "esModuleInterop": true,
        "sourceMap": true,
        "rootDir": "src",
